Revert from med-bio to med-cloud
The `med-bio` metapackage may need user interaction at install time or also pull packages with a large size or a complex dependency chain. So `med-cloud` is more appropriate. Packages that are not installed by `med-cloud` can be installed separately, and of course we can suggest their addition to `med-cloud`'s list.