Remove epoch in Debian package version number.
In Debian packages, an 'epoch' is an optional component of the version
number, comprising everything before the first colon character, if
present.  Epochs are usually not part of the upstream version nuber and
are added to solve distribution-specific problems.

I am using Perl instead of Sed so that I can use a non-greedy regular
expression, just in case the version number after the epoch also has
a colon character.